+++ common/src/main/java/mca/client/gui/BlueprintScreen.java
@@ -38,6 +38,7 @@ public class BlueprintScreen extends ExtendedScreen {
     private final int positionMarriage = 40;
     private Village village;
     private int reputation;
+    private boolean isVillage;
     private Rank rank;
     private Set<String> completedTasks;
     private String page;
@@ -57,6 +58,10 @@ public class BlueprintScreen extends ExtendedScreen {
 
     private Map<Rank, List<Task>> tasks;
     private Map<String, BuildingType> buildingTypes;
+
+    private BuildingType getBuildingType(String type) {
+        return buildingTypes.containsKey(type) ? buildingTypes.get(type) : new BuildingType();
+    }
 
     public BlueprintScreen() {
         super(Text.literal("Blueprint"));
@@ -197,9 +202,11 @@ public class BlueprintScreen extends ExtendedScreen {
                 by += 22 * 3;
 
                 //rename village
-                addDrawableChild(new ButtonWidget(bx, by, 96, 20, Text.translatable("gui.blueprint.renameVillage"), (b) -> {
-                    setPage("rename");
-                }));
+                if (isVillage) {
+                    addDrawableChild(new ButtonWidget(bx, by, 96, 20, Text.translatable("gui.blueprint.renameVillage"), (b) -> {
+                        setPage("rename");
+                    }));
+                }
                 by += 22;
             case "map":
                 //add building
@@ -350,7 +357,11 @@ public class BlueprintScreen extends ExtendedScreen {
         //name
         transform.push();
         transform.scale(2.0f, 2.0f, 2.0f);
-        drawCenteredText(transform, textRenderer, village.getName(), width / 4, height / 4 - 48, 0xffffffff);
+        if (isVillage) {
+            drawCenteredText(transform, textRenderer, village.getName(), width / 4, height / 4 - 48, 0xffffffff);
+        } else {
+            drawCenteredText(transform, textRenderer, new TranslatableText("gui.blueprint.settlement"), width / 4, height / 4 - 48, 0xffffffff);
+        }
         transform.pop();
     }
 
@@ -395,7 +406,7 @@ public class BlueprintScreen extends ExtendedScreen {
         //buildings
         List<Building> hoverBuildings = new LinkedList<>();
         for (Building building : village.getBuildings().values()) {
-            BuildingType bt = buildingTypes.get(building.getType());
+            BuildingType bt = getBuildingType(building.getType());
 
             if (bt.isIcon()) {
                 BlockPos c = building.getCenter();
@@ -454,7 +465,7 @@ public class BlueprintScreen extends ExtendedScreen {
         List<Text> lines = new LinkedList<>();
 
         //name
-        BuildingType bt = buildingTypes.get(hoverBuilding.getType());
+        BuildingType bt = getBuildingType(hoverBuilding.getType());
         lines.add(Text.translatable("buildingType." + bt.name()));
 
         //size
@@ -630,9 +641,10 @@ public class BlueprintScreen extends ExtendedScreen {
         }
     }
 
-    public void setRank(Rank rank, int reputation, Set<String> completedTasks, Map<Rank, List<Task>> tasks, Map<String, BuildingType> buildingTypes) {
+    public void setRank(Rank rank, int reputation, boolean isVillage, Set<String> completedTasks, Map<Rank, List<Task>> tasks, Map<String, BuildingType> buildingTypes) {
         this.rank = rank;
         this.reputation = reputation;
+        this.isVillage = isVillage;
         this.completedTasks = completedTasks;
         this.tasks = tasks;
         this.buildingTypes = buildingTypes;