+++ common/src/main/java/mca/network/s2c/SkinListResponse.java
@@ -5,9 +5,13 @@ import mca.cobalt.network.Message;
 import mca.resources.ClothingList;
 import mca.resources.HairList;
 
+import java.io.Serial;
 import java.util.HashMap;
 
 public class SkinListResponse implements Message {
+    @Serial
+    private static final long serialVersionUID = 3523559818338225910L;
+
     private final HashMap<String, ClothingList.Clothing> clothing;
     private final HashMap<String, HairList.Hair> hair;