+++ fabric/src/main/java/mca/fabric/MCAFabricClient.java
@@ -21,6 +21,7 @@ import mca.item.SirbenBabyItem;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
+import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.fabricmc.fabric.api.client.particle.v1.ParticleFactoryRegistry;
 import net.fabricmc.fabric.api.client.rendering.v1.BlockEntityRendererRegistry;
 import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
@@ -75,9 +76,11 @@ public final class MCAFabricClient extends ClientProxyAbstractImpl implements Cl
                 SirbenBabyItem.hasBeenInvalidated(stack) ? 1 : 0
         );
 
-        BlockRenderLayerMap.INSTANCE.putBlock(BlocksMCA.INFERNAL_FLAME.get(), RenderLayer.getCutout());
+        ClientPlayConnectionEvents.JOIN.register((handler, sender, server) ->
+                MCAClient.onLogin()
+        );
 
-        MCAClient.onInitializeClient();
+        BlockRenderLayerMap.INSTANCE.putBlock(BlocksMCA.INFERNAL_FLAME.get(), RenderLayer.getCutout());
 
         ClientTickEvents.START_CLIENT_TICK.register(MCAClient::tickClient);
     }